Выносит функцию отрисовки событий в днях в отдельную функцию
